fix(managed-asc): validate and verify obfuscation updates
Исправлена обработка ASC/обфускации для managed WireGuard-серверов.

Что изменено:
- frontend-валидация ASC вынесена в отдельный utility и теперь показывает все ошибочные поля сразу, а не только первое;
- backend-валидация ASC также стала агрегирующей и возвращает полный список некорректных параметров;
- ручное отключение обфускации через all-zero ASC теперь идёт через отдельный clear-path, а не через обычную отправку нулевых значений;
- после изменения ASC backend выполняет read-back verification и не возвращает success, если роутер фактически не применил параметры;
- `I1-I5` сохраняются локально только после успешного применения и проверки ASC, а при отключении ASC очищаются;
- экспорт/backup теперь использует общую backend-валидацию ASC;
- исправлена сборка под 32-bit MIPS: генерация больших H-баз переведена с `int` на `int64`;
- добавлены frontend/backend тесты на агрегированную валидацию, отключение ASC, read-back mismatch и mipsle-safe генерацию H-параметров.

Главный результат: больше нет ложного “успешно сохранено”, когда ASC-параметры на роутере фактически не изменились, и пользователь видит полный список проблемных полей при некорректном вводе.

+++ internal/managed/asc_generate.go
@@ -16,11 +16,11 @@ func rndInt(r *rand.Rand, a, b int) int {
 }
 
 func generateHRanges(r *rand.Rand) [4]string {
-	bases := [4]int{100_000_000, 1_200_000_000, 2_400_000_000, 3_600_000_000}
+	bases := [4]int64{100_000_000, 1_200_000_000, 2_400_000_000, 3_600_000_000}
 	var out [4]string
 	for i, base := range bases {
-		offset := rndInt(r, 0, 4_000_000)
-		spread := rndInt(r, 100_000, 500_000)
+		offset := int64(rndInt(r, 0, 4_000_000))
+		spread := int64(rndInt(r, 100_000, 500_000))
 		start := base + offset
 		out[i] = fmt.Sprintf("%d-%d", start, start+spread)
 	}
@@ -28,10 +28,10 @@ func generateHRanges(r *rand.Rand) [4]string {
 }
 
 func generateHSingle(r *rand.Rand) [4]string {
-	bases := [4]int{100_000_000, 1_200_000_000, 2_400_000_000, 3_600_000_000}
+	bases := [4]int64{100_000_000, 1_200_000_000, 2_400_000_000, 3_600_000_000}
 	var out [4]string
 	for i, base := range bases {
-		out[i] = fmt.Sprintf("%d", base+rndInt(r, 0, 4_000_000))
+		out[i] = fmt.Sprintf("%d", base+int64(rndInt(r, 0, 4_000_000)))
 	}
 	return out
 }
